Move predict_function to base OD model
